<refactor>: модульность файло с точкой входа main.js
